# PaperlessServer/Paperless.apimock/wwwroot/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Paperlessk",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/search/autocomplete" : {
      "get" : {
        "operationId" : "AutoComplete",
        "parameters" : [ {
          "explode" : true,
          "in" : "query",
          "name" : "term",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Search" ]
      }
    },
...
  },
...
}